+++ src/runtime/assemblymanager.cs
@@ -49,7 +49,7 @@ namespace Python.Runtime
         /// </summary>
         internal static void Initialize()
         {
-            assemblies = new AssemblyList(16);
+            assemblies = new ConcurrentQueue<Assembly>();
             pypath = new List<string>(16);
 
             AppDomain domain = AppDomain.CurrentDomain;
@@ -138,7 +138,7 @@ namespace Python.Runtime
         internal static void UpdatePath()
         {
             IntPtr list = Runtime.PySys_GetObject("path");
-            int count = Runtime.PyList_Size(list);
+            var count = Runtime.PyList_Size(list);
             if (count != pypath.Count)
             {
                 pypath.Clear();